+++ UserException.php
@@ -278,15 +278,17 @@ class UserException extends \yii\base\UserException {
 				$fileLog = 'All details registered in database with error code '. $this->errorCode .' (errorID '. $errorID .').';
 
 				// Delete expired errors once per session
-				$session = Yii::$app->session;
-				if ($session && !Yii::$app->request->isConsoleRequest && !Yii::$app->response->isSent) {  //will get message if run in CLI: "session_set_cookie_params(): Cannot change session cookie parameters when headers already sent"
-					try {
-						if (!$session->get('wsErrorsCleared')) {
-							$dbConn->createCommand("DELETE FROM `". $databaseTable ."` WHERE err_expire_days IS NOT NULL AND TO_DAYS(err_timestamp) + err_expire_days < TO_DAYS(NOW())")->execute();
-							$session->set('wsErrorsCleared', 1);
+				if (Yii::$app->has('session')) {
+					$session = Yii::$app->session;
+					if ($session && !Yii::$app->request->isConsoleRequest && !Yii::$app->response->isSent) {  //will get message if run in CLI: "session_set_cookie_params(): Cannot change session cookie parameters when headers already sent"
+						try {
+							if (!$session->get('wsErrorsCleared')) {
+								$dbConn->createCommand("DELETE FROM `". $databaseTable ."` WHERE err_expire_days IS NOT NULL AND TO_DAYS(err_timestamp) + err_expire_days < TO_DAYS(NOW())")->execute();
+								$session->set('wsErrorsCleared', 1);
+							}
+						} catch (\Exception $e) {
+							throw $e;
 						}
-					} catch (\Exception $e) {
-						throw $e;
 					}
 				}
 			} else {
